unify the cli and the server into one package
Restructure server to use cobra,
split host and port in configuration
reorganize imports
