+++ caches/redis_cacher_test.go
@@ -1,9 +1,10 @@
 package caches_test
 
 import (
-	"github.com/nanopack/shaman/caches"
 	"testing"
 	"time"
+
+	"github.com/nanopack/shaman/caches"
 )
 
 func initializeRedisCacher(t *testing.T, expires int) caches.Cacher {
