--- cli/config/config.go
@@ -1,12 +1,0 @@
-package config
-
-var (
-	AuthToken string
-	Host      string
-	Port      int
-	Insecure  bool
-)
-
-func init() {
-
-}
