--- cli/main.go
@@ -1,13 +1,0 @@
-package main
-
-// TODO:
-//  - everything
-//  - test
-
-import (
-	"github.com/nanopack/shaman/cli/commands"
-)
-
-func main() {
-	commands.ShamanCli.Execute()
-}
